requirements.txt
anyio==3.6.1
asgiref==3.5.2
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.4
docopt==0.6.2
fastapi==0.78.0
gunicorn==20.1.0
h11==0.13.0
httptools==0.4.0
idna==3.3
numpy==1.22.3
opencv-contrib-python==4.5.5.64
Pillow==9.1.1
pipreqs==0.4.11
pydantic==1.9.0
python-dotenv==0.20.0
python-multipart==0.0.5
PyYAML==6.0
requests==2.28.2
six==1.16.0
sniffio==1.2.0
starlette==0.19.1
typing_extensions==4.2.0
urllib3==1.26.14
uvicorn==0.17.6
watchgod==0.8.2
websockets==10.3
yarg==0.1.9
